backend/requirements.txt
asgiref==3.9.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
Django==4.2.23
django-cors-headers==4.5.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djoser==2.3.1
idna==3.10
oauthlib==3.3.1
pillow==11.3.0
pycparser==2.22
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
